Fix for Version 1.0.9

+++ cszcms/controllers/admin/Plugin_manager.php
@@ -61,26 +61,32 @@ class Plugin_manager extends CI_Controller {
         admin_helper::is_not_admin($this->session->userdata('admin_type'));
         /* upload zip file */
         $zip_ext = array('application/x-zip', 'application/zip', 'application/x-zip-compressed', 'application/s-compressed', 'multipart/x-zip');
-        if (in_array($_FILES['file_upload']['type'], $zip_ext)) {
-            $paramiter = '_1';
-            $photo_id = time();
-            $uploaddir = 'photo/plugin/';
-            $file_f = $_FILES['file_upload']['tmp_name'];
-            $file_name = $_FILES['file_upload']['name'];
-            $upload_file = $this->Csz_admin_model->file_upload($file_f, $file_name, '', $uploaddir, $photo_id, $paramiter);
-            $newfname = FCPATH.$uploaddir.$upload_file;
-        }
-        if (file_exists($newfname)) {
-            @$this->unzip->extract($newfname, FCPATH);
-            if (file_exists(FCPATH . 'plugin_sql/install.sql')) {
-                $this->Csz_admin_model->execSqlFile(FCPATH . 'plugin_sql/install.sql');
-                delete_files(FCPATH . 'plugin_sql', TRUE);
-                rmdir(FCPATH . 'plugin_sql');
+        if($_FILES['file_upload'] != null){
+            if (in_array($_FILES['file_upload']['type'], $zip_ext)) {
+                $paramiter = '_1';
+                $photo_id = time();
+                $uploaddir = 'photo/plugin/';
+                $file_f = $_FILES['file_upload']['tmp_name'];
+                $file_name = $_FILES['file_upload']['name'];
+                $upload_file = $this->Csz_admin_model->file_upload($file_f, $file_name, '', $uploaddir, $photo_id, $paramiter);
+                $newfname = FCPATH.$uploaddir.$upload_file;
+                if (file_exists($newfname)) {
+                    @$this->unzip->extract($newfname, FCPATH);
+                    if (file_exists(FCPATH . 'plugin_sql/install.sql')) {
+                        $this->Csz_admin_model->execSqlFile(FCPATH . 'plugin_sql/install.sql');
+                        delete_files(FCPATH . 'plugin_sql', TRUE);
+                        rmdir(FCPATH . 'plugin_sql');
+                    }
+                    if (is_writable($newfname)) {
+                        @unlink($newfname);
+                    }
+                    $this->session->set_flashdata('error_message', '<div class="alert alert-success" role="alert">' . $this->lang->line('success_message_alert') . '</div>');
+                }else{
+                    $this->session->set_flashdata('error_message','<div class="alert alert-danger" role="alert">'.$this->lang->line('error_message_alert').'</div>');
+                }
+            }else{
+                $this->session->set_flashdata('error_message','<div class="alert alert-danger" role="alert">'.$this->lang->line('pluginmgr_zip_remark').'</div>');
             }
-            if (is_writable($newfname)) {
-                @unlink($newfname);
-            }
-            $this->session->set_flashdata('error_message', '<div class="alert alert-success" role="alert">' . $this->lang->line('success_message_alert') . '</div>');
         }else{
             $this->session->set_flashdata('error_message','<div class="alert alert-danger" role="alert">'.$this->lang->line('error_message_alert').'</div>');
         }
